Add answers to exercise 7 and resulting MIDI files
I named the MIDI files after the sequence of notes generated in each case.

While I initially expected the first file to contain "C D E F G A",
the two sequences concatenated, and the offset to be inserted between
each sequence (relative to the end of the last sequence inserted),
I discovered that the actual file contained only the second sequence.

This helped me to understand that the method seq_notes() inserts all
sequences from the start of the track (the offset is absolute, not relative)
and notes from a new sequence may override notes from a previous sequence.

The sequences that I heard in the two other cases confirmed this analysis.