Add the code needed to start exercise 8, and generated MIDI
The generated MIDI files are random, thus modified at each run.

I noticed a peculiarity in the first sequence, were all notes seem
to be in the same octave, different from one run to the next, but
the same for all notes. This is possible but unlikely by picking
the octave at random from a range.