Use different parameters for the gauss curves of octaves and durations
While the music is melodious and the major scale is well rendered,
the rhythm still sounds odd: very short notes mixed at random with
longer notes blur any idea of tempo.

I will next try to shift the top of the gauss curve to the first
item for durations, to get a very quick tempo with only very few
longer notes.

I will then add a memory in random_notes, giving a high weight to
durations close to the duration of the previous note, whether it was
fast or slow.