Add unit tests for shifted offset
The value used as parameter is the opposite of the value that
makes sense to me: -1 must be provided to place the top of the
gauss curve on the item at index 1.

I will update the definition to make the value compatible with
common sense: +1 to shift the gauss curve to the right, -1 to
shift the gauss curve to the left.