Start to answer exercise 13
This first answer applies a condition too strong for the requirements:
both sets related by transposition and inversion get removed, and we
obtain the 12 sets from 3-1 to 3-12 listed above the exercise in the book.
The condition should be relaxed to keep sets related by inversion.
For example, the set (0,4,6) is related to (0,2.6) (mirrored),
but should be preserved when only transpositions are considered.