Complete exercise 13, using lexical order to order and filter sets
Each chord can be transposed to three different sets (all identical
in the case of chord (0,4,8)) starting with note 0:

* (0,a,b)
* (0,b-a,12-a)
* (0,12-b,12+a-b)

Comparing these three sets using lexical order allows to keep
only the minimum set, skipping the two variations obtained
by transposition.

There are 55 sets starting with 0 among the 220 sets of three notes.
If you set aside (0,4,8) which is identical to itself in each of
the tree transpositions described above, only one in three sets
is kept in the remaining 54 sets, which results in 1 + 54/3 = 19 sets.