Prepare exercise 14: add example code from the book
The script so far runs the unmodified harmonize() methods
of Note and NoteSeq and checks that the values correspond
to those expected (and printed in the book).

In the next step, I will extend the classes Note and NoteSeq,
adding a new method harmonize_template().

I will then add assertions for the results expected when calling
harmonize_template() with various templates, including templates
which correspond to current examples.

Finally, I will add the implementation of harmonize_template()
and replace the implementation of harmonize() to use
harmonize_template() instead, and check that all unit tests pass.